Fixed Bug
I think I have able to locate the bug in the design that is responsible
for dancing table. I could not check the database because my sql is not
working in my PC. I think I have to install XAMP. Anyways check the GUI
and tell me if the bug is fixed or not.